Revert "[core] limit parallelly read file memory usage, extract some methods (#1072)"
This reverts commit 5d23c7da2ae4973c7d70f9e89a4b83220264142c.
